fix: move DataSyncService class into separate file
